Fix Year as a Date on Either Side for Anime
The anime date parser was broken in certain edge cases where the end date
was a plain year. Chronic can't parse it properly as it's ambiguous.
Instead, it was just returning the current date. Now, we do a check for
a bare year after splitting on both the start and end dates to avoid this
problem.